Refactor the code for the netlify functions
- Created the zipRes_Uid function to reuse in various places
